+++ core/app/helpers/spree/base_helper.rb
@@ -4,11 +4,11 @@ module Spree
       text = text ? h(text) : Spree.t(:cart)
       css_class = nil
 
-      if simple_current_order.nil? || simple_current_order.item_count.zero?
+      if current_order.nil? || current_order.item_count.zero?
         text = "#{text}: (#{Spree.t(:empty)})"
         css_class = 'empty'
       else
-        text = "#{text}: (#{simple_current_order.item_count})  <span class='amount'>#{simple_current_order.display_total.to_html}</span>"
+        text = "#{text}: (#{current_order.item_count})  <span class='amount'>#{current_order.display_total.to_html}</span>"
         css_class = 'full'
       end